# Conflicts:
#	src/gameobj/Actor.java
#	src/gameobj/Barrier.java
#	src/gameobj/Building.java
#	src/gameobj/GameObject.java
#	src/gameobj/View.java
#	src/scenes/MainScene.java
#	src/scenes/SecondScene.java
#	src/scenes/StartMenuScene.java
#	src/util/MapGenerator.java

+++ src/util/MapGenerator.java
@@ -50,7 +50,7 @@ public class MapGenerator {
                 Map newMap = this.mapPool.get(poolIndex++);
                 // 設定完 x, y 後再加入，不然預設地圖位置皆為 0, 0
                 newMap.setX((float) Global.MAP_WIDTH * x);
-                newMap.setY((float) (float) Global.MAP_HEIGHT * y);
+                newMap.setY((float) Global.MAP_HEIGHT * y);
                 this.maps.add(newMap);
             }
         }
@@ -68,7 +68,7 @@ public class MapGenerator {
                 Map newMap = this.mapPool.get(poolIndex++);
                 // 設定完 x, y 後再加入，不然預設地圖位置皆為 0, 0
                 newMap.setX((float) Global.MAP_WIDTH * x);
-                newMap.setY((float) (float) Global.MAP_HEIGHT * y);
+                newMap.setY((float) Global.MAP_HEIGHT * y);
                 this.maps.add(newMap);
             }
         }
